+++ packages/client/src/pages/Forum/Table/CreateTopicModal/styles.module.scss
@@ -21,34 +21,3 @@
   gap: 16px;
   justify-content: space-between;
 }
-
-.pagination {
-  margin-top: 2.5rem;
-  overflow: initial !important;
-}
-
-.user {
-  display: flex;
-  align-items: center;
-  justify-content: start;
-
-  .avatar {
-    margin-right: 0.5rem;
-  }
-}
-
-.tr {
-  cursor: pointer;
-}
-
-.tr .tc {
-  border-bottom: 1px solid #fff;
-}
-
-.tc.tc_align_right {
-  text-align: right;
-}
-
-.tc.creator_header {
-  padding-left: calc(56px + 0.5rem);
-}
